Identify model from manufacturer data, not name
The BLE advertisement does not always contain useful data in the name
field. It may, for example, contain a variant of the MAC address
"18-04-ED-61-E4-4D", instead of a product name such as "Wave+".

To handle this situation more robustly, base the Airthings product
identification on the manufacturer data field. This should always be
present in a valid advertisement. For Wave devices, the manufacturer
data decodes to a serial number with the first 4 digits identifying the
specific model, e.g. "2930" for a Airthings Wave+.